requirements.txt
requests
Pillow
python-telegram-bot
playwright
beautifulsoup4
geopy
